requirements.txt
fastapi
python-multipart
uvicorn
jinja2
pydub
google-genai
python-dotenv
aiofiles
audioop-lts